refactor: organize classes into packages
- Fulfills professor's requirement to organize code in folders (packages in Java)
- Created model, storage, ui, and main packages
- Adjusted imports and visibility modifiers across packages while strictly maintaining original file and class names
- Preserved previous error handling, business logic, and documentation changes
