+++ storage/IArmazenador.java
@@ -4,6 +4,10 @@
  * @author Kaua Bezerra, Liam Vedovato, Raul Kolaric, Rodrigo Ward 
  * @version 24/03/2026
  */
+package storage;
+
+import model.Aluno;
+
 public interface IArmazenador {
     /**
      * Insere um novo aluno no armazenamento.
