+++ ui/MenuTexto.java
@@ -1,14 +1,16 @@
 /**
- * Classe que implementa a interface IMenu fornecendo uma interface grafica.
+ * Classe que implementa a interface IMenu fornecendo uma interface de texto via console.
  * 
  * @author Kaua Bezerra, Liam Vedovato, Raul Kolaric, Rodrigo Ward 
  * @version 24/03/2026
  */
-import javax.swing.JOptionPane;
+package ui;
 
-public class MenuGrafico implements IMenu {
+import java.util.Scanner;
+
+public class MenuTexto implements IMenu {
     /**
-     * Cria um menu grafico utilizando JOptionPane.
+     * Cria um menu de texto lendo a entrada do console.
      * 
      * @param opcoes Array de strings com as opcoes do menu.
      * @return A opcao numerica escolhida pelo usuario.
@@ -19,25 +21,26 @@ public class MenuGrafico implements IMenu {
         String itens = "";
         for (int i = 0; i < opcoes.length; i++){
             itens = itens + "\n" + opcoes[i];
+            
         }
         
         itens = itens + "\n\nSelecione a opcao: ";
+
+        Scanner sc = new Scanner(System.in);
         boolean entradaValida = false;
         opcao = 0;
-        
+
         do {
+            System.out.print(itens);
+            String s = sc.nextLine();
             try {
-                String input = JOptionPane.showInputDialog(itens);
-                if (input == null) {
-                    return 6; // Se o usuario cancelar, encerra o loop assumindo 'sair'
-                }
-                opcao = Integer.parseInt(input);
+                opcao = Integer.parseInt(s);
                 entradaValida = true;
             } catch (NumberFormatException e) {
-                JOptionPane.showMessageDialog(null, "Entrada invalida. Por favor, digite um numero valido.");
+                System.out.println("\nEntrada invalida. Por favor, digite um numero valido.\n");
             }
         } while (!entradaValida);
-        
+
         return opcao;
     }
 }
